rename schedule/dialog to scheduleDialog, extract to new component SchedulingFormGroup
