+++ src/web/pages/schedules/ScheduleDialog.jsx
@@ -22,6 +22,7 @@ import TimeZoneSelect from 'web/components/form/TimeZoneSelect';
 import {ComponentWithHoverCard} from 'web/components/hover-card';
 import Row from 'web/components/layout/Row';
 import ManualLink from 'web/components/link/ManualLink';
+import useFeatures from 'web/hooks/useFeatures';
 import useTranslation from 'web/hooks/useTranslation';
 import DaySelect from 'web/pages/schedules/DaySelect';
 import MonthDaysSelect from 'web/pages/schedules/MonthDaysSelect';
@@ -69,6 +70,66 @@ const updateDateWithTime = (currentDate, timeString) => {
     .clone()
     .hour(parsedTime.hour())
     .minute(parsedTime.minute());
+};
+
+const SchedulingFormGroup = ({
+  timezone,
+  startDate,
+  startTime,
+  handleStartDateChange,
+  handleTimeChange,
+  handleNowButtonClick,
+}) => {
+  const [_] = useTranslation();
+  const features = useFeatures();
+
+  const formGroup = (
+    <FormGroup label={_('Scheduling')} title={_('Scheduling')}>
+      <Row align={'end'} flex="row" gap={'lg'}>
+        <DatePicker
+          label={_('Start Date')}
+          name="startDate"
+          timezone={timezone}
+          value={startDate}
+          onChange={handleStartDateChange}
+        />
+        <TimePicker
+          label={_('Start Time')}
+          name="startDate"
+          value={startTime}
+          onChange={newStartTime => handleTimeChange(newStartTime, 'startTime')}
+        />
+        <Button title={_('Now')} onClick={handleNowButtonClick} />
+      </Row>
+    </FormGroup>
+  );
+
+  if (!features.featureEnabled('ENABLE_CONTAINER_SCANNING')) {
+    return formGroup;
+  }
+
+  return (
+    <ComponentWithHoverCard
+      dataTestId="scheduling-options-info"
+      helpAriaLabel={_('More information about scheduling options')}
+      helpContent={
+        <>
+          {_(
+            'Please consider enabling "Time synchronization" if the scans are not starting at the expected time.',
+          )}
+          <div>
+            <ManualLink
+              anchor="configuring-the-time-synchronization"
+              page="managing-gos"
+            >
+              {_('Learn more')}
+            </ManualLink>
+          </div>
+        </>
+      }
+      slot={formGroup}
+    />
+  );
 };
 
 const ScheduleDialog = ({
@@ -425,46 +486,13 @@ const ScheduleDialog = ({
                 onChange={onValueChange}
               />
             </FormGroup>
-            <ComponentWithHoverCard
-              dataTestId="scheduling-options-info"
-              helpAriaLabel={_('More information about scheduling options')}
-              helpContent={
-                <>
-                  {_(
-                    'Please consider enabling "Time synchronization" if the scans are not starting at the expected time.',
-                  )}
-                  <div>
-                    <ManualLink
-                      anchor="configuring-the-time-synchronization"
-                      page="managing-gos"
-                    >
-                      {_('Learn more')}
-                    </ManualLink>
-                  </div>
-                </>
-              }
-              slot={
-                <FormGroup title={_('Scheduling')}>
-                  <Row align={'end'} flex="row" gap={'lg'}>
-                    <DatePicker
-                      label={_('Start Date')}
-                      name="startDate"
-                      timezone={timezone}
-                      value={startDate}
-                      onChange={handleStartDateChange}
-                    />
-                    <TimePicker
-                      label={_('Start Time')}
-                      name="startDate"
-                      value={startTime}
-                      onChange={newStartTime =>
-                        handleTimeChange(newStartTime, 'startTime')
-                      }
-                    />
-                    <Button title={_('Now')} onClick={handleNowButtonClick} />
-                  </Row>
-                </FormGroup>
-              }
+            <SchedulingFormGroup
+              handleNowButtonClick={handleNowButtonClick}
+              handleStartDateChange={handleStartDateChange}
+              handleTimeChange={handleTimeChange}
+              startDate={startDate}
+              startTime={startTime}
+              timezone={timezone}
             />
 
             <FormGroup title={_('Timezone')}>
