+++ src/web/pages/schedules/__tests__/Dialog.test.jsx
@@ -15,7 +15,7 @@ import {
 import date from 'gmp/models/date';
 import Schedule from 'gmp/models/schedule';
 import timezones from 'gmp/time-zones';
-import ScheduleDialog from 'web/pages/schedules/Dialog';
+import ScheduleDialog from 'web/pages/schedules/ScheduleDialog';
 
 const checkElementVisibilityAndContent = (
   labelText,
